pyproject.toml
[project]
name = "eda-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "google-genai>=1.29.0",
    "grandalf>=0.8",
    "graphviz>=0.21",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.4",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-prebuilt>=0.6.4",
    "matplotlib>=3.10.5",
    "mlflow>=3.2.0",
    "msgpack>=1.1.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = []
